added sending of raw odometry, bugfix 3

+++ msl_tmc_motion/src/CNMCPacket.cpp
@@ -129,12 +129,12 @@ namespace msl_driver
 	//**** CONVERT METHODS ****//
 	short CNMCPacket::convertByteToShort(int start)
 	{
-		unsigned short s = 0;
-		s = ((unsigned short) this->data->at(start)) << 8;
-		s += (unsigned short) this->data->at(start+1);
+		short s = 0;
+		s = ((short) this->data->at(start)) << 8;
+		s += (short) this->data->at(start+1);
 //		return (short)(((unsigned short)(this->data->at(start)) << 8 ) + (unsigned char)this->data->at(start+1));
 
-		return (short) s;
+		return s;
 	}
 
 	std::vector<uint8_t> CNMCPacket::convertShortToByte(short data)
